api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend for DAHack AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",
    "tortoise-orm>=0.20.0",
    "aerich>=0.7.2",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0.post1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "testcontainers>=3.7.1",
    "asyncpg>=0.29.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"
# asyncio_default_fixture_loop_scope = "function"


[dependency-groups]
dev = ["pyright>=1.1.401", "pytest-asyncio>=1.0.0", "ruff>=0.11.12"]
